+++ CHANGELOG.md
@@ -9,6 +9,7 @@ under development
 - `Content.SetContentTypeJSON` and `Content.SetContentTypeXML` append `charset=utf-8` to content type header.
 - Add `Context.HTML` to send HTML response.
 - Add `Context.Cookie` and `Context.Cookies`.
+- Add `Context.FormValue`.
 
 v1.8.1 April 2, 2020
 --------------------